--- README.md
@@ -1,40 +1,0 @@
-Материалы третьей лекции:
-=======
-TBD
-
-Домашнее задание:
-=======
-Вам надо разбиться в команды по два-три человека. Если вы не смогли найти себе команду - то как можно быстрее сообщите нам об этом.
-
-Каждая команда должна сделать форк (один, чтобы потом можно было его сдать в виде пулл-реквеста) и вести весь проект в этом репозитории. После создания репозитория добавить права на push для другого участника можно в разделе Admin->Collaborators вашего проекта.
-
-Программа должна показывать на экране поле для ввода слова на английском и кнопку "Перевести", по которой она должна найти в интернете перевод этого слова и 10 картинок по этому слову и показать их на втором экране. По клавише "Back" программа должна возвращаться к первому экрану.
-
-Программа делится на две части:
-
-1) Весь user interface - поле ввода, логика перехода между экранами, перевод (en->ru) и его отображение на втором экране.
-2) Поиск в интернете по английскому слову десяти картинок про это слово, скачивание и отображение.
-
-Команда должна решить, кто какую из этих двух частей пишет.
-
-Пожалуйста, не забывайте про файл .gitignore и не добавляйте лишних файлов в репозиторий.
-
-Полезная информация для выполнения ДЗ:
-=======
- - Activity, Intent, startActivity() - для вызова второго экрана (не забывайте объявлять вторую Activity в манифесте!).
- - putExtra(), getStringExtra() - для передачи данных между экранами.
- - http://api.yandex.ru/translate/ - один из вариантов API для перевода (можно пользоваться любым).
- - GridView, ListView, Gallery - варианты layout-ов для отрисовки нескольких картинок сразу.
- - AsyncTask - для неблокирующей (асинхронной) загрузки данных.
-
-
-Порядок сдачи:
-=======
-Сдавать задание нужно в виде форка и пулл-реквеста к https://github.com/IFMO-MobDev-2014/lesson3, в описании укажите ФИО и номера групп.
-Подробнее про пулл-реквесты можно почитать тут, например: http://habrahabr.ru/post/125999/ и https://help.github.com/articles/using-pull-requests.
-
-Результат принимается до понедельника (30 сентября) 23:59. После этого оценка за это домашнее задание автоматически снижается в два раза. Напоминаем, что хотя задание дается и команде, оценивается оно лично.
-
-Оценки:
-=======
-http://bit.ly/ifmo-md-14-results
