+++ Translator/.gitignore
@@ -1,0 +1,5 @@
+.gradle
+/local.properties
+/.idea/workspace.xml
+.DS_Store
+/build
