+++ Translator/app/.gitignore
