+++ Translator/app/src/main/res/layout/activity_edit.xml
@@ -1,0 +1,31 @@
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    tools:context=".EditActivity">
+
+    <EditText
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/editField"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true"
+        android:autoText="true"
+        android:clickable="true"
+        android:textSize="@dimen/text_size"
+        android:visibility="visible"
+        android:text="@string/edit_information"
+        android:layout_below="@+id/backButton" />
+
+    <Button
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/tr_but_name"
+        android:id="@+id/backButton"
+        android:layout_alignParentTop="false"
+        android:layout_centerHorizontal="true" />
+</RelativeLayout>
