+++ Translator/app/src/main/res/values/styles.xml
