+++ Translator/gradle.properties
