+++ Translator/gradlew.bat
