+++ Translator/settings.gradle
