--- app/src/main/java/com/example/artur/lesson3/StrToImages.java
@@ -1,6 +1,0 @@
-package com.example.artur.lesson3;
-
-
-public class StrToImages {
-    //yoyoyo
-}
